Refactor last-harvest-timestamp-report
