+++ src/job/job-functions-map.ts
@@ -2,7 +2,7 @@ import { generateReportsDaily as countResourcesJobFunction } from "../job-functi
 import { generateReportsDaily as checkSessionJobFunction } from "../job-functions/generate-session-complete-report-daily/index.js";
 import { DataMonitoringFunction } from "../types.js";
 import { dummyFunction } from "./dummy.js";
-import { getHarvestingTimestampDaily } from "./harvesting-timestamp-daily.js";
+import { getHarvestingTimestampDaily } from "../job-functions/generate-last-harvest-timestamp-report-daily/index.js";
 import { JobFunction } from "./job.js";
 
 // Effectively an index of some kind. Maps the enum values of DataMonitoringFunction to the data monitoring functions of type 'JobFunction'.
